% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSunlightTimes.R
\name{getSunlightTimes}
\alias{getSunlightTimes}
\title{Get Sunlight times}
\usage{
getSunlightTimes(date = NULL, lat = NULL, lon = NULL, data = NULL,
  keep = c("solarNoon", "nadir", "sunrise", "sunset", "sunriseEnd",
  "sunsetStart", "dawn", "dusk", "nauticalDawn", "nauticalDusk", "nightEnd",
  "night", "goldenHourEnd", "goldenHour"), tz = "UTC")
}
\arguments{
\item{date}{: \code{Date}. Single or multiple Date. YYYY-MM-DD}

\item{lat}{: \code{numeric}. Single latitude}

\item{lon}{: \code{numeric}. Single longitude}

\item{data}{: \code{data.frame}. Alternative to use \code{date}, \code{lat}, \code{lon} for passing multiple coordinates}

\item{keep}{: \code{character}. Vector of variables to keep. See \code{Details}}

\item{tz}{: \code{character}. timezone of results}
}
\value{
\code{data.frame}
}
\description{
Get Sunlight times
}
\details{
Availabled variables are :

\itemize{
  \item{"sunrise"}{ : sunrise (top edge of the sun appears on the horizon)}
  \item{"sunriseEnd"}{ : sunrise ends (bottom edge of the sun touches the horizon)}
  \item{"goldenHourEnd"}{ : morning golden hour (soft light, best time for photography) ends}
  \item{"solarNoon"}{ : solar noon (sun is in the highest position)}
  \item{"goldenHour"}{ : evening golden hour starts}
  \item{"sunsetStart"}{ : sunset starts (bottom edge of the sun touches the horizon)}
  \item{"sunset"}{ : sunset (sun disappears below the horizon, evening civil twilight starts)}
  \item{"dusk"}{ : dusk (evening nautical twilight starts)}
  \item{"nauticalDusk"}{ : nautical dusk (evening astronomical twilight starts)}
  \item{"night"}{ : night starts (dark enough for astronomical observations)}
  \item{"nadir"}{ : nadir (darkest moment of the night, sun is in the lowest position)}
  \item{"nightEnd"}{ : night ends (morning astronomical twilight starts)}
  \item{"nauticalDawn"}{ : nautical dawn (morning nautical twilight starts)}
  \item{"dawn"}{ : dawn (morning nautical twilight ends, morning civil twilight starts)}
}
}
\examples{

# one date
getSunlightTimes(date = Sys.Date(), lat = 50.1, lon = 1.83, tz = "CET")

# multiple date + subset
getSunlightTimes(date = seq.Date(Sys.Date()-9, Sys.Date(), by = 1), 
  keep = c("sunrise", "sunriseEnd", "sunset", "sunsetStart"), 
  lat = 50.1, lon = 1.83, tz = "CET")
  
# multiple coordinates
data <- data.frame(date = seq.Date(Sys.Date()-9, Sys.Date(), by = 1), 
    lat = c(rep(50.1, 10), rep(49, 10)), 
    lon = c(rep(1.83, 10), rep(2, 10)))
    
getSunlightTimes(data = data, 
  keep = c("sunrise", "sunriseEnd", "sunset", "sunsetStart"), tz = "CET")
      
}
